+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/EquoIdeGradlePlugin.java
@@ -69,10 +69,6 @@ public class EquoIdeGradlePlugin implements Plugin<Project> {
 								task -> {
 									task.setGroup(TASK_GROUP);
 									task.setDescription("Launches an Eclipse-based IDE for this project");
-
-									task.getCaching().set(caching);
-									task.getMavenDeps().set(equoIde);
-									task.getProjectDir().set(project.getProjectDir());
 								});
 		project.afterEvaluate(
 				unused -> {
@@ -88,8 +84,12 @@ public class EquoIdeGradlePlugin implements Plugin<Project> {
 										});
 						equoIdeTask.configure(
 								task -> {
+									task.getCaching().set(caching);
 									task.getQuery().set(query);
+									task.getMavenDeps().set(equoIde);
+									task.getProjectDir().set(project.getProjectDir());
 									task.getUseAtomos().set(extension.useAtomos);
+									task.ideHooks = extension.getIdeHooks();
 								});
 					} catch (Exception e) {
 						throw new RuntimeException(e);