+++ solstice/src/main/java/dev/equo/ide/BuildPluginIdeMain.java
@@ -11,16 +11,25 @@
  * Contributors:
  *     EquoTech, Inc. - initial API and implementation
  *******************************************************************************/
-package dev.equo.solstice;
+package dev.equo.ide;
 
+import com.diffplug.common.swt.os.OS;
+import dev.equo.solstice.NestedJars;
+import dev.equo.solstice.SerializableMisc;
+import dev.equo.solstice.Solstice;
+import dev.equo.solstice.SolsticeInit;
+import dev.equo.solstice.SolsticeManifest;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
+import java.util.Objects;
+import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Stream;
+import org.eclipse.swt.widgets.Display;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.BundleException;
 import org.osgi.framework.InvalidSyntaxException;
@@ -31,6 +40,94 @@ import org.osgi.framework.InvalidSyntaxException;
  * in other contexts as well.
  */
 public class BuildPluginIdeMain {
+	public static class Caller {
+		public IdeLockFile lockFile;
+		public IdeHook.List ideHooks;
+		public File workspaceDir;
+		public ArrayList<File> classpath;
+		public BuildPluginIdeMain.DebugClasspath debugClasspath;
+		public Boolean initOnly, showConsole, useAtomos;
+		public String showConsoleFlag;
+
+		public void launch() throws IOException, InterruptedException {
+			Objects.requireNonNull(lockFile);
+			Objects.requireNonNull(ideHooks);
+			Objects.requireNonNull(workspaceDir);
+			Objects.requireNonNull(classpath);
+			Objects.requireNonNull(debugClasspath);
+			Objects.requireNonNull(initOnly);
+			Objects.requireNonNull(showConsole);
+			Objects.requireNonNull(useAtomos);
+			Objects.requireNonNull(showConsoleFlag);
+
+			var nestedJarFolder = new File(workspaceDir, NestedJars.DIR);
+			for (var nested : NestedJars.inFiles(classpath).extractAllNestedJars(nestedJarFolder)) {
+				classpath.add(nested.getValue());
+			}
+
+			var ideHooksFile = new File(workspaceDir, "ide-hooks");
+			var ideHooksCopy = ideHooks.copy();
+			ideHooksCopy.add(IdeHookLockFile.forWorkspaceDir(workspaceDir));
+			SerializableMisc.toFile(ideHooksCopy, ideHooksFile);
+
+			var classpathSorted = Launcher.copyAndSortClasspath(classpath);
+			debugClasspath.printWithHead(
+					"jars about to be launched", classpathSorted.stream().map(File::getAbsolutePath));
+			boolean isBlocking =
+					initOnly || showConsole || debugClasspath != BuildPluginIdeMain.DebugClasspath.disabled;
+			var vmArgs = new ArrayList<String>();
+			if (OS.getRunning().isMac()) {
+				vmArgs.add("-XstartOnFirstThread");
+			}
+			vmArgs.add("-Dorg.slf4j.simpleLogger.defaultLogLevel=" + (isBlocking ? "info" : "error"));
+
+			Consumer<Process> monitorProcess;
+			if (isBlocking) {
+				monitorProcess = null;
+			} else {
+				// if we're spawning a new IDE, record the lockfile before it launches
+				long lockFileBeforeLaunch = lockFile.read();
+				monitorProcess =
+						process -> {
+							// sleep over and over until the lockfile changes
+							while (lockFile.read() == lockFileBeforeLaunch) {
+								try {
+									Thread.sleep(10);
+								} catch (InterruptedException e) {
+									// ignore
+								}
+							}
+							// kill the console that we've been waiting on as a solution to
+							// https://github.com/equodev/equo-ide/issues/44
+							process.destroyForcibly();
+						};
+			}
+			var exitCode =
+					Launcher.launchJavaBlocking(
+							isBlocking,
+							classpathSorted,
+							vmArgs,
+							BuildPluginIdeMain.class.getName(),
+							monitorProcess,
+							"-installDir",
+							workspaceDir.getAbsolutePath(),
+							"-useAtomos",
+							Boolean.toString(useAtomos),
+							"-initOnly",
+							Boolean.toString(initOnly),
+							"-debugClasspath",
+							debugClasspath.name(),
+							"-ideHooks",
+							ideHooksFile.getAbsolutePath());
+			if (!isBlocking) {
+				System.out.println("NEED HELP? If the IDE doesn't appear, try adding " + showConsoleFlag);
+			}
+			if (exitCode != 0) {
+				System.out.println("WARNING! Exit code: " + exitCode);
+			}
+		}
+	}
+
 	public enum DebugClasspath {
 		disabled,
 		names,
@@ -108,8 +205,22 @@ public class BuildPluginIdeMain {
 		boolean initOnly = parseArg(args, "-initOnly", Boolean::parseBoolean, false);
 		DebugClasspath debugClasspath =
 				parseArg(args, "-debugClasspath", DebugClasspath::valueOf, DebugClasspath.disabled);
+		File hookListFile = parseArg(args, "-ideHooks", File::new, null);
+		IdeHook.List ideHooksParsed;
+		if (hookListFile == null) {
+			ideHooksParsed = new IdeHook.List();
+		} else {
+			ideHooksParsed = SerializableMisc.fromFile(IdeHook.List.class, hookListFile);
+		}
 		debugClasspath.printAndExitIfEnabled();
 
+		IdeHook.InstantiatedList ideHooks = ideHooksParsed.instantiate();
+		if (!initOnly) {
+			ideHooks.forEach(IdeHookInstantiated::beforeDisplay);
+			var display = Display.getDefault();
+			ideHooks.forEach(IdeHookInstantiated::afterDisplay, display);
+			ideHooks.forEach(IdeHookInstantiated::beforeOsgi);
+		}
 		BundleContext context;
 		if (useAtomos) {
 			// the spelled-out package is on purpose so that Atomos can remain an optional component
@@ -119,6 +230,10 @@ public class BuildPluginIdeMain {
 		} else {
 			context = Solstice.initialize(new SolsticeInit(installDir));
 		}
+		if (!initOnly) {
+			ideHooks.forEach(IdeHookInstantiated::afterOsgi, context);
+		}
+
 		if (initOnly) {
 			System.out.println(
 					"Loaded "
@@ -129,7 +244,7 @@ public class BuildPluginIdeMain {
 			return;
 		}
 
-		int exitCode = IdeMainUi.main(context);
+		int exitCode = IdeMainUi.main(context, ideHooks);
 		if (exitCode == 0) {
 			System.exit(0);
 		} else {